+++ src/main/java/wooteco/subway/line/LineController.java
@@ -3,6 +3,8 @@ package wooteco.subway.line;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.LineResponse;
 
 import java.net.URI;
 import java.util.List;
