+++ src/main/java/wooteco/subway/station/StationController.java
@@ -3,6 +3,8 @@ package wooteco.subway.station;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
 
 import java.net.URI;
 import java.util.List;
